refactor fodler structure
move errroBoduanry to its won folder
move errroPage to its own folder
update imports
add backgrond image to errorPage
