+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import './App.css';
 
-import ErrorBoundary from './components/errorBoundary';
+import ErrorBoundary from './components/error-boundary/errorBoundary';
 import HomePage from './pages/home/home.page';
 
 function App() {
