+++ client/src/pages/home/error/error.page.jsx
@@ -3,10 +3,10 @@ import React from 'react';
 function ErrorPage({ error }) {
     console.log(error);
     return (
-        <article className="errorArticle">
-            {
-                JSON.stringify(error)
-            }
+        <article
+            className="errorArticle"
+            style={{backgroundImage: "url('../assets/error.png')"}}
+            >
         </article>
     );
 }
